+++ lib/src/commonMain/kotlin/co/touchlab/kjwt/cryptography/ext/JwkExt.kt
@@ -1,4 +1,4 @@
-package co.touchlab.kjwt.ext
+package co.touchlab.kjwt.cryptography.ext
 
 import co.touchlab.kjwt.annotations.ExperimentalKJWTApi
 import co.touchlab.kjwt.cryptography.toCryptographyKotlin
