+++ lib/src/commonTest/kotlin/co/touchlab/kjwt/ClaimsValidationTest.kt
@@ -1,5 +1,7 @@
 package co.touchlab.kjwt
 
+import co.touchlab.kjwt.cryptography.ext.signWith
+import co.touchlab.kjwt.cryptography.ext.verifyWith
 import co.touchlab.kjwt.exception.ExpiredJwtException
 import co.touchlab.kjwt.exception.IncorrectClaimException
 import co.touchlab.kjwt.exception.MissingClaimException
