+++ lib/src/commonTest/kotlin/co/touchlab/kjwt/JweDecodeTest.kt
@@ -2,6 +2,8 @@
 
 package co.touchlab.kjwt
 
+import co.touchlab.kjwt.cryptography.ext.decryptWith
+import co.touchlab.kjwt.cryptography.ext.encryptWith
 import co.touchlab.kjwt.exception.ExpiredJwtException
 import co.touchlab.kjwt.exception.IncorrectClaimException
 import co.touchlab.kjwt.exception.MalformedJwtException
