+++ lib/src/commonTest/kotlin/co/touchlab/kjwt/JwtBuilderHeaderTest.kt
@@ -1,5 +1,9 @@
 package co.touchlab.kjwt
 
+import co.touchlab.kjwt.cryptography.ext.decryptWith
+import co.touchlab.kjwt.cryptography.ext.encryptWith
+import co.touchlab.kjwt.cryptography.ext.signWith
+import co.touchlab.kjwt.cryptography.ext.verifyWith
 import co.touchlab.kjwt.ext.contentType
 import co.touchlab.kjwt.ext.getHeader
 import co.touchlab.kjwt.ext.getHeaderOrNull
