+++ lib/src/commonTest/kotlin/co/touchlab/kjwt/JwtKeyRegistryTest.kt
@@ -1,24 +1,31 @@
 package co.touchlab.kjwt
 
+import co.touchlab.kjwt.annotations.DelicateKJWTApi
+import co.touchlab.kjwt.cryptography.ext.decryptWith
+import co.touchlab.kjwt.cryptography.ext.encryptWith
+import co.touchlab.kjwt.cryptography.ext.registerEncryptionKey
+import co.touchlab.kjwt.cryptography.ext.registerSigningKey
+import co.touchlab.kjwt.cryptography.ext.signWith
+import co.touchlab.kjwt.cryptography.ext.verifyWith
 import co.touchlab.kjwt.ext.subjectOrNull
 import co.touchlab.kjwt.model.algorithm.EncryptionAlgorithm
 import co.touchlab.kjwt.model.algorithm.EncryptionContentAlgorithm
 import co.touchlab.kjwt.model.algorithm.SigningAlgorithm
 import co.touchlab.kjwt.cryptography.registry.EncryptionKey
-import co.touchlab.kjwt.cryptography.registry.CryptographyKotlinJwtKeyRegistry
 import co.touchlab.kjwt.cryptography.registry.SigningKey
+import co.touchlab.kjwt.model.registry.DefaultJwtKeyRegistry
 import io.kotest.core.spec.style.FunSpec
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 
-class JwtKeyRegistryTest :
+@OptIn(DelicateKJWTApi::class) class JwtKeyRegistryTest :
     FunSpec({
 
         context("sign using registry") {
 
             test("sign HS256 using registry signing key") {
                 val key = hs256Key()
-                val registry = CryptographyKotlinJwtKeyRegistry()
+                val registry = DefaultJwtKeyRegistry()
                 registry.registerSigningKey(
                     SigningKey.SigningOnlyKey(
                         identifier = SigningKey.Identifier(SigningAlgorithm.HS256, null),
@@ -45,7 +52,7 @@ class JwtKeyRegistryTest :
 
             test("sign HS256 using registry with kid") {
                 val key = hs256Key()
-                val registry = CryptographyKotlinJwtKeyRegistry()
+                val registry = DefaultJwtKeyRegistry()
                 registry.registerSigningKey(
                     SigningKey.SigningOnlyKey(
                         identifier = SigningKey.Identifier(SigningAlgorithm.HS256, "sign-key"),
@@ -71,7 +78,7 @@ class JwtKeyRegistryTest :
             }
 
             test("sign throws when no matching key in registry") {
-                val registry = CryptographyKotlinJwtKeyRegistry()
+                val registry = DefaultJwtKeyRegistry()
 
                 assertFailsWith<IllegalArgumentException> {
                     Jwt
@@ -86,7 +93,7 @@ class JwtKeyRegistryTest :
 
             test("parser delegates verification to shared registry") {
                 val key = hs256Key()
-                val sharedRegistry = CryptographyKotlinJwtKeyRegistry()
+                val sharedRegistry = DefaultJwtKeyRegistry()
                 sharedRegistry.registerSigningKey(
                     SigningKey.VerifyOnlyKey(
                         identifier = SigningKey.Identifier(SigningAlgorithm.HS256, null),
@@ -113,7 +120,7 @@ class JwtKeyRegistryTest :
 
             test("parser delegates verification to shared registry with kid") {
                 val key = hs256Key()
-                val sharedRegistry = CryptographyKotlinJwtKeyRegistry()
+                val sharedRegistry = DefaultJwtKeyRegistry()
                 sharedRegistry.registerSigningKey(
                     SigningKey.VerifyOnlyKey(
                         identifier = SigningKey.Identifier(SigningAlgorithm.HS256, "k1"),
@@ -145,7 +152,7 @@ class JwtKeyRegistryTest :
                         dev.whyoleg.cryptography.algorithms.SHA256,
                         "wrong-secret-at-least-256-bits-long-padding".encodeToByteArray(),
                     )
-                val sharedRegistry = CryptographyKotlinJwtKeyRegistry()
+                val sharedRegistry = DefaultJwtKeyRegistry()
                 sharedRegistry.registerSigningKey(
                     SigningKey.VerifyOnlyKey(
                         identifier = SigningKey.Identifier(SigningAlgorithm.HS256, null),
@@ -177,7 +184,7 @@ class JwtKeyRegistryTest :
 
             test("encrypt Dir A256GCM using registry encryption key") {
                 val cek = aesSimpleKey(256)
-                val registry = CryptographyKotlinJwtKeyRegistry()
+                val registry = DefaultJwtKeyRegistry()
                 registry.registerEncryptionKey(
                     EncryptionKey.EncryptionOnlyKey(
                         identifier = EncryptionKey.Identifier(EncryptionAlgorithm.Dir, null),
@@ -204,7 +211,7 @@ class JwtKeyRegistryTest :
 
             test("encrypt Dir A256GCM using registry with kid") {
                 val cek = aesSimpleKey(256)
-                val registry = CryptographyKotlinJwtKeyRegistry()
+                val registry = DefaultJwtKeyRegistry()
                 registry.registerEncryptionKey(
                     EncryptionKey.EncryptionOnlyKey(
                         identifier = EncryptionKey.Identifier(EncryptionAlgorithm.Dir, "enc-k1"),
@@ -230,7 +237,7 @@ class JwtKeyRegistryTest :
             }
 
             test("encrypt throws when no matching key in registry") {
-                val registry = CryptographyKotlinJwtKeyRegistry()
+                val registry = DefaultJwtKeyRegistry()
 
                 assertFailsWith<IllegalArgumentException> {
                     Jwt
@@ -245,7 +252,7 @@ class JwtKeyRegistryTest :
 
             test("parser delegates decryption to shared registry") {
                 val cek = aesSimpleKey(256)
-                val sharedRegistry = CryptographyKotlinJwtKeyRegistry()
+                val sharedRegistry = DefaultJwtKeyRegistry()
                 sharedRegistry.registerEncryptionKey(
                     EncryptionKey.DecryptionOnlyKey(
                         identifier = EncryptionKey.Identifier(EncryptionAlgorithm.Dir, null),
@@ -272,7 +279,7 @@ class JwtKeyRegistryTest :
 
             test("parser delegates decryption to shared registry with kid") {
                 val cek = aesSimpleKey(256)
-                val sharedRegistry = CryptographyKotlinJwtKeyRegistry()
+                val sharedRegistry = DefaultJwtKeyRegistry()
                 sharedRegistry.registerEncryptionKey(
                     EncryptionKey.DecryptionOnlyKey(
                         identifier = EncryptionKey.Identifier(EncryptionAlgorithm.Dir, "enc-k1"),
@@ -301,15 +308,15 @@ class JwtKeyRegistryTest :
         context("delegation cycle detection") {
 
             test("self-delegation throws") {
-                val registry = CryptographyKotlinJwtKeyRegistry()
+                val registry = DefaultJwtKeyRegistry()
                 assertFailsWith<IllegalArgumentException> {
                     registry.delegateTo(registry)
                 }
             }
 
             test("direct cycle throws") {
-                val a = CryptographyKotlinJwtKeyRegistry()
-                val b = CryptographyKotlinJwtKeyRegistry()
+                val a = DefaultJwtKeyRegistry()
+                val b = DefaultJwtKeyRegistry()
                 a.delegateTo(b)
                 assertFailsWith<IllegalArgumentException> {
                     b.delegateTo(a)
@@ -317,9 +324,9 @@ class JwtKeyRegistryTest :
             }
 
             test("transitive cycle throws") {
-                val a = CryptographyKotlinJwtKeyRegistry()
-                val b = CryptographyKotlinJwtKeyRegistry()
-                val c = CryptographyKotlinJwtKeyRegistry()
+                val a = DefaultJwtKeyRegistry()
+                val b = DefaultJwtKeyRegistry()
+                val c = DefaultJwtKeyRegistry()
                 a.delegateTo(b)
                 b.delegateTo(c)
                 assertFailsWith<IllegalArgumentException> {
@@ -328,9 +335,9 @@ class JwtKeyRegistryTest :
             }
 
             test("linear chain without cycle is allowed") {
-                val a = CryptographyKotlinJwtKeyRegistry()
-                val b = CryptographyKotlinJwtKeyRegistry()
-                val c = CryptographyKotlinJwtKeyRegistry()
+                val a = DefaultJwtKeyRegistry()
+                val b = DefaultJwtKeyRegistry()
+                val c = DefaultJwtKeyRegistry()
                 a.delegateTo(b)
                 b.delegateTo(c)
                 // no exception expected
@@ -341,7 +348,7 @@ class JwtKeyRegistryTest :
 
             test("sign and verify using a registry with merged SigningKeyPair") {
                 val key = hs256Key()
-                val sharedRegistry = CryptographyKotlinJwtKeyRegistry()
+                val sharedRegistry = DefaultJwtKeyRegistry()
                 // Registering complementary keys merges them into a SigningKeyPair
                 sharedRegistry.registerSigningKey(
                     SigningKey.SigningOnlyKey(
@@ -375,7 +382,7 @@ class JwtKeyRegistryTest :
 
             test("encrypt and decrypt using a registry with merged EncryptionKeyPair") {
                 val cek = aesSimpleKey(256)
-                val sharedRegistry = CryptographyKotlinJwtKeyRegistry()
+                val sharedRegistry = DefaultJwtKeyRegistry()
                 // Register DecryptionOnlyKey first, then EncryptionOnlyKey — merges into EncryptionKeyPair
                 sharedRegistry.registerEncryptionKey(
                     EncryptionKey.DecryptionOnlyKey(
