+++ lib/src/commonTest/kotlin/co/touchlab/kjwt/TestFixtures.kt
@@ -1,9 +1,9 @@
 package co.touchlab.kjwt
 
 import co.touchlab.kjwt.cryptography.SimpleKey
-import co.touchlab.kjwt.ext.key
-import co.touchlab.kjwt.ext.newKey
-import co.touchlab.kjwt.ext.parse
+import co.touchlab.kjwt.cryptography.ext.key
+import co.touchlab.kjwt.cryptography.ext.newKey
+import co.touchlab.kjwt.cryptography.ext.parse
 import co.touchlab.kjwt.model.algorithm.EncryptionAlgorithm
 import co.touchlab.kjwt.model.algorithm.SigningAlgorithm
 import co.touchlab.kjwt.cryptography.registry.EncryptionKey
